UT-3585: fix cyclic dependency
