+++ bin/common/common.py
@@ -1,0 +1,33 @@
+#!/usr/bin/env python3
+
+import os
+
+from typing import List
+
+
+def remove_file(file_path: str):
+    if os.path.isfile(file_path):
+        os.remove(file_path)
+
+
+def is_conversion_success(status_code: int, file_path: str):
+    return os.path.exists(file_path) and status_code == 0
+
+
+def remove_tmp_and_exit(status_code: int, file_path: str):
+    remove_file(file_path)
+    if status_code != 0:
+        return 1
+
+
+def run_command_and_convert_to_pdfa(command: List[str], tmp_path: str, target_path: str):
+    initial_convert_status_code = run_shell_command(command)[0]
+    if not is_conversion_success(initial_convert_status_code, tmp_path):
+        remove_tmp_and_exit(initial_convert_status_code, tmp_path)
+
+    status_code = pdf2pdfa(tmp_path, target_path)
+    if not is_conversion_success(status_code, target_path):
+        remove_tmp_and_exit(status_code, tmp_path)
+
+    remove_file(tmp_path)
+    return 0
